--- src/main/java/team/washer/server/v2/domain/smartthings/service/DetectMachineCompletionService.java
@@ -1,8 +1,0 @@
-package team.washer.server.v2.domain.smartthings.service;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-public interface DetectMachineCompletionService {
-    Optional<LocalDateTime> execute(String deviceId);
-}
