--- src/main/java/team/washer/server/v2/domain/smartthings/service/DetectMachinePausedService.java
@@ -1,5 +1,0 @@
-package team.washer.server.v2.domain.smartthings.service;
-
-public interface DetectMachinePausedService {
-    boolean execute(String deviceId);
-}
