--- src/main/java/team/washer/server/v2/domain/smartthings/service/QueryDeviceStatusService.java
@@ -1,7 +1,0 @@
-package team.washer.server.v2.domain.smartthings.service;
-
-import team.washer.server.v2.domain.smartthings.dto.response.SmartThingsDeviceStatusResDto;
-
-public interface QueryDeviceStatusService {
-    SmartThingsDeviceStatusResDto execute(String deviceId);
-}
