--- src/main/java/team/washer/server/v2/domain/smartthings/service/impl/DetectMachineCompletionServiceImpl.java
@@ -1,45 +1,0 @@
-package team.washer.server.v2.domain.smartthings.service.impl;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import team.washer.server.v2.domain.smartthings.service.DetectMachineCompletionService;
-import team.washer.server.v2.domain.smartthings.service.QueryDeviceStatusService;
-import team.washer.server.v2.global.util.DateTimeUtil;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class DetectMachineCompletionServiceImpl implements DetectMachineCompletionService {
-
-    private final QueryDeviceStatusService queryDeviceStatusService;
-
-    @Override
-    @Transactional(readOnly = true)
-    public Optional<LocalDateTime> execute(String deviceId) {
-        try {
-            var status = queryDeviceStatusService.execute(deviceId);
-
-            var washerJobState = status.getWasherJobState();
-            var dryerJobState = status.getDryerJobState();
-
-            // 세탁기: "finish", 건조기: "finished" (SmartThings 명세 기준)
-            var isFinished = "finish".equalsIgnoreCase(washerJobState) || "finished".equalsIgnoreCase(dryerJobState);
-
-            if (isFinished) {
-                log.debug("Device {} job is finished", deviceId);
-                return Optional.ofNullable(DateTimeUtil.parseAndConvertToKoreaTime(status.getCompletionTime()));
-            }
-
-            return Optional.empty();
-        } catch (Exception e) {
-            log.warn("Failed to detect completion state for device: {}", deviceId, e);
-            return Optional.empty();
-        }
-    }
-}
