--- src/main/java/team/washer/server/v2/domain/smartthings/service/impl/QueryDeviceStatusServiceImpl.java
@@ -1,44 +1,0 @@
-package team.washer.server.v2.domain.smartthings.service.impl;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import team.themoment.sdk.exception.ExpectedException;
-import team.washer.server.v2.domain.smartthings.dto.response.SmartThingsDeviceStatusResDto;
-import team.washer.server.v2.domain.smartthings.repository.SmartThingsTokenRepository;
-import team.washer.server.v2.domain.smartthings.service.QueryDeviceStatusService;
-import team.washer.server.v2.global.thirdparty.smartthings.feign.SmartThingsFeignClient;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class QueryDeviceStatusServiceImpl implements QueryDeviceStatusService {
-
-    private final SmartThingsFeignClient feignClient;
-    private final SmartThingsTokenRepository tokenRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public SmartThingsDeviceStatusResDto execute(String deviceId) {
-        try {
-            var token = tokenRepository.findSingletonToken()
-                    .orElseThrow(() -> new ExpectedException("SmartThings 토큰이 존재하지 않습니다", HttpStatus.NOT_FOUND));
-
-            if (!token.isValid()) {
-                throw new ExpectedException("SmartThings 토큰이 만료되었거나 유효하지 않습니다", HttpStatus.NOT_FOUND);
-            }
-
-            var authorization = "Bearer " + token.getAccessToken();
-            return feignClient.getDeviceStatus(authorization, deviceId);
-        } catch (ExpectedException e) {
-            log.error("smartthings token not found or invalid", e);
-            throw e;
-        } catch (Exception e) {
-            log.error("smartthings failed to query device status deviceId={}", deviceId, e);
-            throw new ExpectedException("기기 상태 조회에 실패했습니다: " + e.getMessage(), HttpStatus.BAD_GATEWAY);
-        }
-    }
-}
